# dist/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Pick_IRegionType.Exclude_keyofIRegionType.regionId-or-status__": {
				"properties": {
					"designation": {
						"type": "string"
					},
					"etat": {
						"type": "string"
					}
				},
				"required": [
					"designation",
					"etat"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IRegionType.regionId-or-status_": {
				"$ref": "#/components/schemas/Pick_IRegionType.Exclude_keyofIRegionType.regionId-or-status__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Pick_IRegionType.Exclude_keyofIRegionType.status__": {
				"properties": {
					"designation": {
						"type": "string"
					},
					"regionId": {
						"type": "string"
					},
					"etat": {
						"type": "string"
					}
				},
				"required": [
					"designation",
					"etat"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IRegionType.status_": {
				"$ref": "#/components/schemas/Pick_IRegionType.Exclude_keyofIRegionType.status__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Pick_IRegionType.regionId_": {
				"properties": {
					"regionId": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
...
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "uptodate-ekivalApp-api",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/admin/regions": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"city": {},
										"status": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"city",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"501": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Region"
				],
				"security": [
					{
						"Bearer": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_IRegionType.regionId-or-status_"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "Etats",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"city": {},
										"status": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"city",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Region"
				],
				"security": [
					{
						"Bearer": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 100,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"city": {},
										"status": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"city",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"501": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Region"
				],
				"security": [
					{
						"Bearer": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_IRegionType.status_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {},
										"status": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					},
					"501": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Region"
				],
				"security": [
					{
						"Bearer": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Pick_IRegionType.regionId_"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/v1"
		}
	]
}